* commit '5833e60e466cbd5425094062152088cf5aa79cd7':
  changing storage from double to int.

+++ utility-generator/src/main/java/eu/melodic/upperware/utilitygenerator/evaluator/EvaluatingUtils.java
@@ -114,7 +114,7 @@ public class EvaluatingUtils {
 
                 //real
                 case STORAGE:
-                    predicates.add(getStoragePredicate((double) var.getValue())); //fixme - to check
+                    predicates.add(getStoragePredicate((int) var.getValue())); //fixme - to check
                     break;
                 case LOCATION:
                     break;
